chore(global): move package for add error docs feature
