+++ src/main/java/online/lifeasgame/global/tracing/MDCKeys.java
@@ -1,4 +1,4 @@
-package online.lifeasgame.global.constant;
+package online.lifeasgame.global.tracing;
 
 public final class MDCKeys {
     private MDCKeys() {}
